+++ internal/provider/storage/storage_lun_data_source.go
@@ -76,6 +76,13 @@ type StorageLunDataSourceSpaceModel struct {
 	Size       types.Int64 `tfsdk:"size"`
 	Used       types.Int64 `tfsdk:"used"`
 	Allocation types.Bool  `tfsdk:"scsi_thin_provisioning_support_enabled"`
+	Guarantee *StorageLunDataSourceReservationModel  `tfsdk:"guarantee"`
+}
+
+// StorageLunDataSourceReservationModel describes the data source data model for space guarantee.
+type StorageLunDataSourceReservationModel struct {
+	Requested types.Bool `tfsdk:"requested"`
+	Reserved types.Bool `tfsdk:"reserved"`
 }
 
 // Metadata returns the data source type name.
@@ -167,6 +174,20 @@ func (d *StorageLunDataSource) Schema(ctx context.Context, req datasource.Schema
 						MarkdownDescription: "Specifies the value for the space allocation attribute, which determines if the LUN supports the SCSI Thin Provisioning features",
 						Computed:            true,
 					},
+					"guarantee": schema.SingleNestedAttribute{
+						MarkdownDescription: "Specifies the space guarantee state for the LUN",
+						Computed:            true,
+						Attributes: map[string]schema.Attribute{
+							"requested": schema.BoolAttribute{
+								MarkdownDescription: "Indicates if the space is requested",
+								Computed:            true,
+							},
+							"reserved": schema.BoolAttribute{
+								MarkdownDescription: "Indicates if the space is reserved",
+								Computed:            true,
+							},
+						},
+					},
 				},
 			},
 			"id": schema.StringAttribute{
@@ -235,6 +256,10 @@ func (d *StorageLunDataSource) Read(ctx context.Context, req datasource.ReadRequ
 		Size:       types.Int64Value(restInfo.Space.Size),
 		Used:       types.Int64Value(restInfo.Space.Used),
 		Allocation: types.BoolPointerValue(restInfo.Space.Allocation),
+		Guarantee: &StorageLunDataSourceReservationModel{
+			Requested: types.BoolPointerValue(restInfo.Space.Guarantee.Requested),
+			Reserved:  types.BoolPointerValue(restInfo.Space.Guarantee.Reserved),
+		},
 	}
 	data.ID = types.StringValue(restInfo.UUID)